Move customers data inline in App.js
